+++ packages/server/src/instances/sandbox/service.ts
@@ -4,14 +4,18 @@ import { Logger, Service } from '@botpress/messaging-engine'
 import { validate as uuidValidate } from 'uuid'
 import { ClientService } from '../../clients/service'
 import { MappingService } from '../../mapping/service'
-import { InstanceService } from '../service'
+import { InstanceMessagingService } from '../messaging/service'
 
 const JOIN_KEYWORD = '!join'
 
 export class InstanceSandboxService extends Service {
   private logger = new Logger('Sandbox')
 
-  constructor(private clients: ClientService, private mapping: MappingService, private instances: InstanceService) {
+  constructor(
+    private clients: ClientService,
+    private mapping: MappingService,
+    private messaging: InstanceMessagingService
+  ) {
     super()
   }
 
@@ -40,7 +44,7 @@ export class InstanceSandboxService extends Service {
   }
 
   async printAskJoinSandbox(conduitId: uuid, endpoint: Endpoint) {
-    await this.instances.sendToEndpoint(conduitId, endpoint, {
+    await this.messaging.send(conduitId, endpoint, {
       type: 'text',
       text: 'Please join the sandbox by sending : !join your_passphrase'
     })
@@ -64,7 +68,7 @@ export class InstanceSandboxService extends Service {
   }
 
   async printWrongPassphrase(conduitId: uuid, endpoint: Endpoint) {
-    await this.instances.sendToEndpoint(conduitId, endpoint, {
+    await this.messaging.send(conduitId, endpoint, {
       type: 'text',
       text: 'Wrong passphrase'
     })
